Defined the proof-of-concept for field traits.
Defined 3 field traits for the 3 kinds of access(shared/mutable/by value),
which take in the name of the field as a generic parameter
(currently encoded as a tuple of unit structs).

Defined the GetFieldExt trait as the way to access fields generically,
delegating to the 3 field traits(when required),
defining additional methods on top of the other field traits
(like the methods for accessing multiple fields mutably).