Changed GetField to be object safe,changed `GetFieldExt::field_mut_*` to be infallible.
Removed NAME and INDEX associated constants from GetField to make it object safe..

The `GetFieldExt::field_mut_*` methods take in the `MultiTString` type,
which guarantees that the TStrings it contains are not repeated inside.

The MultiTString is constructed using the `rstr!()` macro,
which ensures that none of its string parameters are equal to each other.