Added `#[struc(no_trait)]` helper attribute.
This attribute disables the generation of the `<deriving_type>_SI` trait,
which aliases all the accessor traits for the deriving type,
and has a blanket impl for any type with at least the same accessor trait impls.

Also added an example in docs for `#[derive(Structural)]` using this helper attribute.