Start of 0.2 rewrite commit chunking,with rewrite of proc macros.
This is the first commit of many,
each commit will include a few of the changes since the commit from a week ago,
and they may not compile until the last of the changes is committed.

Changes in structural_derive:

Removed TStringSet.

Replaced `TStringSet` with `FieldPaths`
(along with FieldPath,FieldPathComponent,PathUniqueness),
to generate all code including the `FieldPath` or `FieldPathSet` types,

Changed `tident_tokenizer` to `tident_tokens`,returning a `proc_macro2::TokenStream`
instead of an `impl ToTokens`

Added `StructPmr` variant to `FullPathForChars`,
for outputing TString types in a context that imports `structural::pmr as __struct_pmr`.

Replaced TI and ti macros with FP and fp macros,
outputting a `FieldPath`/`FieldPathSet` instead of a `TString`/`TStringSet`.

Removed `StrOrIdent`,renaming the `str_or_ident` module to `ident_or_index`.

Added 1 test case for the `FP` macro in derive crate.