Fixes and additional tests,improved structural root mod example.
Fixed implementation of specialized functions.

Changed the default field access to mutable and by value,
equivalent to the `#[struc(access="mut move")]` attribute on a field.

Fixed parsing of float literals in `fp` and `FP` ending with a `.`,
interpreting it as separating the next field path component.

Removed `[]` and `()` syntax from `fp` and `FP`  macros.

Wrote more tests for `fp` and `FP` macros