Implemented macros for enums.
Fixed Rev* traits to use the combined errors of all intermediate field accessors.

Tweaked existing macros to support enums with an `Err` variant
(which conflicts with the GetFieldImpl::Err associated type if written as `Self::Err`).

Added the `z_unsafe_impl_get_field_raw_mut_method_enum` macro to implement the
unsafe methods in GetFieldMutImpl.

Added the `impl_getter_enum` and `impl_getters_for_derive_enum` macros,
to implement Structural and the field accessor traits for enums.

Added impls of the field accessor traits for Result and Option(with tests for Option only).

Defined the `MakeUnit` trait to get `()` and pointers to it in generic contexts.