Rewrite to allow every operation on enum trait objects,won't compile for a few commits.
Before the rewrite enum trait objects could only:
- Check whether it is a variant (whose name is passed as a parameter).
- Access a variant field optionally.

After the rewrite it will also be possible to:
- Do an unchecked access of a variant field.
- Construct a proxy for a variant to access multiple non-optional fields

Starting with rewrite of derive macro crate.

Added a way to output the `*VariantField*` trait ident from the access and optionality.

Added syntax for getting the VariantProxy for a variant,
written as `::Variant` without any field.
This outputs the yet uncommited `VariantName<_>` type.

Changed generation of variant and field identifiers to output a TString,
manually wrapping it in the appropriate type in the generated code.