Defined the EnumExt trait,made IsVariant unsafe,defined more field path types.
The EnumExt trait provides extensio methods to wrap Self in VariantProxy
if it is the variant whose name is passed as a parameter.

Made the IsVariant trait unsafe,whose safety invariant is relied on by VariantProxy.

Added the from_box constructor to construct a `Box<VariantProxy<T,V>>` from a `Box<T>`.

Rewrote VariantProxy accessor impls due to the rewrite.

Defined the VariantName marker type,to represent the name of a variant in a field path.

Defined the UncheckedVariantField marker type,
to access a variant field without checking that an enum is that variant.

Implemented Copy+Clone and constructors for VariantField / VariantName / UncheckedVariantField.

Removed the MakeUnit trait because it's unused.