Rewrote a few macros,defined a few new ones.
Renamed the delegation macro to unsafe_delegate_structural_with.
Updated its implementation based on the changes to the accessor traits,
and to delegate the `IsVariant` and `*VariantField*` traits.
Added more options for how the raw pointer accessor method gets delegated.

Rewrote the enum derivation macros to the new way of accessing enum fields.

Removed the accessor impls for enums that return VariantProxy,
since this is done in the `Rev*Field` impls for VariantName,
and because it can't be both object safe and return `&VariantProxy<Self,_>`.

Defined the `unsafe_delegate_variant_field` macro to delegate the impls of the
`IsVariant` and `*VariantField*` traits.

Defined the `map_fe` macro.

Defined the `try_optionality`/`map_optionality` macros,
which behave differently based on the optionality of a field.

Defined the `vf_err` macro to compute the error type of a
field accessor impls that takes `UncheckedVariantField` as a parameter.