Added impls for slices(&[T],&mut [T],Box<[T]>,Rc<[T]>,Arc<[T]>).
Added ToUsize trait to convert a type-level value to a usize,
used for example to get a 100usize from a `fp!(100)` to index a slice.

Added `#[doc(hidden)]` ToDigit trait to convert type-level characters to a u8.

Added IsBit trait to convert Bit0 and Bit1 to bool.

Added IsStructural trait,and implemented it for every structural type,
This trait allows `&` to delegate to a `dyn Trait` (which implements IsStructural),
as well as have a separate impl for `&[T]`.

Made `IsStructural` a supertrait of GetFieldExt.