Improved support of newtype variants.
Now Structural generates an additional structural alias for structs
that can be used as the bound for newtype variants containing the struct.
The structural alias is called `DerivingType_VSI`,
and takes the name of the variant as an additional type parameter.

Added the `#[struc(replace_bounds="Trait<@variant>")]` attribute for enum variants,
to replace the bound for the fields of the enum variant with the ones passed to the attribute.
`@variant` arguments are replaced with `TStr!( Foo )`,
where Foo is the name of the variant that the attribute was used on.

Fixed bug where `Opt*VariantField*` trait bounds were being outputted for
non-optional enum fields.

Added checks that attributes for the generated trait error when the trait is not generated.

Replaces testing of bounds with the Privacies1_SI trait+impl with the
`assert_equal_bounds` macro.

Added optional fields to `variants_with_accesses::Foo` to test that
optional fields in structural aliases work correctly.

Defined test for newtype variants with replaced bounds.

Tweaked the docs for optional accessors.