Added optional arguments to `#[struc(delegate_to())]`.
The arguments are:
- `bound="T:bound"`: Adds the constraint to all the trait impls.
- `mut_bound="T:bound"`: Adds the constraint to the `GetFieldImpl` impl.
- `into_bound="T:bound"`: Adds the constraint to the `IntoFieldImpl` impl.

Added an examples that passes the `*bound` arguments.

Added calls to mutable getter methods in enum tests.

Changed the default value of the argument for `raw_mut_impl`
(part of the input to the delegation macro) from `?Sized` to `Sized`

Added test for `#[struc(delegate_to())]` with bounds,
using the newly defined `declare_querying_trait` macro to test whether a trait
is implemented or not.

Implemented Deref for VariantProxy.

Changed VariantProxy to implement GetFieldMutImpl for `?Sized` types,
which uses specialization for efficiency.

Changed VariantProxy::from_box and EnumExt::box_into_variant
to return a VariantProxy<Box<T>,_> instead of a Box<VariantProxy<T,_>>.

Added a DelegateTo struct in structural_derive::structural_derive::delegation
to store the parsed arguments to `#[struc(delegate_to())]`.

Ran `cargo fmt`