Starting commits for adding exhaustive enums and other things.
This will not compile until the commit says so.

Added the VariantCount trait to specify the amount of variants in an enum.

Added the machinery for the `switch` macro to work with exhaustive and nonexhaustive enums.

Renaming `TString` to `TStr_`,not everywhere until this compiles.

Added IsTStr to assert that a type is a `TStr_`.

Added the Debug+MarkerType bounds to IsFieldPath.
Added the Debug bound to IsFieldPathSet.

Added and improved docs for UncheckedVariantField.