Added hidden `_Structural_BorrowSelf` trait for autoreferencing Self in macros.
Fixed all warnings for enum derivation macro,
and errors from using `map_optionality` without prefixing with `$crate::`.

Reexporting `_switch_tstring_aliases` for later use in `switch` macro.

Defined the GetRefKind trait and `RefKind` enum,
to test the kind of reference of `Self`,this is to get around `&mut` coercing to `&`.