Rewrote major parts of the `switch` macro.
The switch macro can now destructure variant fields after it's been matched.
It supports:
- copying Copy fields with the `&field_name` or `&mut field_name` syntax.
- Ignoring a field with `_`,not requiring the field to exist(this is used by tuple variants).

Moved most of the implementation of `switch` into the `#[doc(hidden)]``switch_inn` macro,
this is to have a fallthrough case in `switch` when no access mode is specified.

Added access modes(similar to default match bindings),which determine how variants are destructured,
specifying the default access mode for all variants in the switch header.

Fixed support for `switch!{self;...}` ,
since the `self` identifier cannot be reused for the variant proxy.

Improved and modified the docs for the macro to reflect all the changes made to it.