Made Option a regular enum,replaced `IsFieldPath*` traits with more generic ones.
Now Option is a regular enum with a transparent Some variant and an None variant,
requiring `.Some` or `::Some` to get the value out of the Some variant.

Replaced IsFieldPath trait with the IsSingleFieldPath trait.

Replaced IsFieldPathSet trait with the IsMultiFieldPath trait.

Implemented IsSingleFieldPath for all structural types that implement RevGetField.

Implemented IsMultiFieldPath for all structural types that implement RevGetMultiField.

Added examples for `GetFieldType2`,`GetFieldType3`,and `GetFieldType4`

Added docs for `RevGetMultiField*` traits

Added IsFieldErr bounds for the GetFieldImpl::Err and RevGetField::Err associated types

Moved the Ty associated type from RevGetField to the newly created RevFieldType,
to not require a lifetime parameter just to query the type of a field.

Added IsSingleFieldPath as a supertrait of RevFieldType.

Added IsMultiFieldPath as a supertrait of RevGetMultiField.

Renamed NestedFieldSet to NestedFieldPathSet for clarity
(makes it more obvious that it's a field path).