Made TSstr_ public.
Declared the structural::P::TS type,
`#[doc(hidden)]` to add eventually add a cargo feature to use a `&'static str`
const parameter instead of a type.

Moved the type-level chars to the newly created structural::p module.

Defined the pub(crate) TStrPriv type alias to operate on the type-level characters
inside a TStr_.

Moved the definition and most impls of field path components to
structural::field_path::path_components private submodule
(publically reexported to the structural::field_path module).

Updated all the links to IsTStr to point to TStr_

Rewrote some (but not all) code that implicitly used TStr_ as a parameter to
explicitly do so.

Added #[doc(hidden)] reexports of field path types in the root module so that
compiler errors are shorter
(compiler errors print the path that the code uses on my machine)

Changed derive crate to always use the full path to the type-level characters inside a TStr_.