Changed what FP!() and fp!() produce,to make for shorter error messages.
Renamed TStr_ to TStr.

Now `FP!()`/`fp!()` produce 6 possible types:
TStr,VariantField,Variantname,FieldPAth,FieldPathSet,NestedFieldPathSet.

Changed the accessor trait type parameter from FieldPath to TStr,and VariantField,
It's still possible to pass the type that `FP!` evaluates to,
because `FP!(hello)` evaluates to a `TStr` and `FP!(::Foo.bar)` evaluates to a `VariantName`.

Changed everything that used to take a `FieldPath<TStr<_>>` to take a `TStr<_>`.

Fixed invocation of `fp!(::A.0.0)` (it gets passed `:: A . 0.0`)