Fixed error message for specialized trait impls,
Fixed it by moving specialization to a separate impl that is delegated to.
This fix made it so that `this.fields_mut(fp!(::Foo.bar))` prints:
    `T does not implement GetFieldMutImpl<VariantField<TS!(Foo),TS!(bar)>>`,
where before it would say:
    `VariantField<TS!(Foo),TS!(bar)>> does not implement RevGetFieldMut<'_,T>`
    `T does not implement GetFieldImpl<VariantField<TS!(Foo),TS!(bar)>>>`

Declared SpecGetFieldMut and SpecRevGetFieldMut to use in specialization impls.

Moved example enums to for_examples module,and added more enums for examples.

Adeded docs to TStr for how to get a TStr type (in a type context)