Fixed every relative link to work on stable Rust.
For some reason `cargo +nightly doc` has an always-enabled
feature in nightly that translates [foo] to a link to the `foo` item,
which doesn't work on stable.

This is the only nightly "feature" I've seen in Rust that doesn't require an
explicit opt-in compared to the same version of the stable compiler,
so I found it really surprising that it's nightly only.