Added docs to Structural impl.
Extracted the logic for code generation from the structural alias macro implementation
into a the new `write_docs` module.

Added docs for the outputted `Structural` trait impl detailing the accessor traits
that were implemented.
This is not complete,it generates the wrong docs for newtype variants,
and no docs for types that delegate to a field.

Added `#[struc(no_docs)]` attribute in both Structural and structural_alias,
which disables docs in generated code.

Added tests for the generated code with documentation enabled and disabled.