Improved generated docs for traits.
Removed names module,always outputting  `TStr<....>` directly.

Changed VariantIdent so that it's used as the name of a StructuralVariant,
and it can be either a concrete `TStr` parameter or a generic parameter in
the generated code.

Updated doc generation to handle:
- generic variant names.
- newtype variants.

Changed docs for `*_VSI` traits to list its fields inside a variant with a generic name.

Added docs to all `*_ESI` traits.

Fixed a bug where the rustdoc generated docs show `Foo_VSI<TStr<TS<"Foo_VSI<@variant>">>>`
as a bound when it didn't exist in the generated code.
Now it would be `Foo_VSI<TStr<TS<"VariantName">>>`.