Bumped minimum Rust version to 1.40
Bumped the minimum required Rust version to 1.40 to
remove the space separated character syntax for the `fp` `FP` `ts` and `TS` macros.
By bumping the MSRV,the documentation examples can use the full FP and TS macros,
including writing identifiers (eg:`FP!(foo)`) and everything else that's valid for
the constructor equivalents of those macros(`fp` and `ts`).

Removed the "rust_1_36","rust_1_40","better_macros",and "nightly_better_macros" features,
always enabling the items cfg-ed out using those features

Simplified the implementation of the `fp` `FP` `ts` and `TS` macros a bit.

Updated the `_TStr_from_ident` macro when const generic are enabled to use stringify,
passing the `stringify!($ident)` to a separate macro as an expression (not a literal),
it's not allowed as a const argument without passing it to another macro first.