Improved aliasing field path detection.
Changed aliasing field path detection to treat `::foo`  and `.foo` the same,
this is to ensure that additions to field path syntax/semantics aren't unsound,
and implementations of `GetField` can alias `GetVariantField`.

Added tests ensuring that many FieldPathSet cases are considered aliasing.

Documented the new behavior in the docs for the `fp` macro.