Added `DropFields::pre_move` method,updated everything accordingly.
The `#[struc(pre_move="foo")]` attribute was defined to delegate the implemetation of
the method to the passed in function.Also added docs for the attribute.

Every wrapper type was updated to use this where necessary.

Every macro was updated to support it.

The tests that use PrePostDropFields now also test the `#[struc(pre_move="foo")]` attribute.

The documetation for how dropping works was updated.

Replaced the doc example for `PrePostDropFields` with one for `DropFields::pre_move`,
to show how to do Drop-like behavior in safe code (`PrePostDropFields` is an unsafe trait).

Defined the RunOnDrop helper type,to run code on Drop.
This type also allows reborrowing the value
(the value is passed to a closure when RunOnDrop is Dropped).