More doc improvements,fixed bugs,tweaked API.
Added docs and example for DerefNested.

Reexporting the DerefNested and DerefNestedOut items in structural::utils
and made the `deref_nested` submodule private.

Renamed ffi_panic_message to abort_with_info and hid it from the docs.

Made all methods in the `field::ownership::on_drop` module unsafe to call,
with the exception of `IntoFieldsWrapper::new`.

Fixed a few bugs that made tests not compile or pass.